requirement.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.16.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.31
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.1
python-decouple==3.6
pytz==2022.5
PyYAML==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
